Add LineStreamer for executable programs
Create a very, very basic streamer for executable programs, using an
interface similar to that of the Slack streamer.

This does sort of work, but it's very low-quality. There's a lot of
missing error handling and a lot of other cases to be considered before
this should go anywhere near production use.

For example, when I connect "ed" to Slack using this and type "q",
there's an extra "?" before it terminates. What am I sending there?
Plus, I need to auto-restart the program.

I'm wondering if it would be cleaner to hide some of the channel bits
behind an io.Reader, and just use that as cmd.Stdin... could be
interesting.