Update README and comments
Update the README to add a bug I want to fix, but also remove a section
that I don't think is relevant. That is, I don't think that LineStreamer
could be a ReadWriter unless there's a well-defined way to do
multi-select over ReadWriters. Or maybe I don't need multi-select at all
and stuff could just happen across multiple goroutines? In any case I'm
just going to stick with the LineStreamer concept for now.

I also put in some implementation notes for the execLineStreamer, and
how I might want to improve that.