Rely on SIGCHLD to terminate slackbridge
Use the SIGCHLD signal from the exiting child program to terminate
slackbridge, rather than rely on the Wait call. This feels slightly
dirty, but there seems to be some circular dependency between a
goroutine in the exec package wanting to read from Slack and us trying
to close slackio. I think I might just need to put the operations in a
better order.