Close slackio before calling Wait
Ensure that slackio is closed *before* Wait gets called. This ensures
that slackio has a chance to close its internal pipes, which is required
to stop the exec package's internal goroutines and terminate slackbridge
when the child exits.

Notably, a pipe I/O error is occurring somewhere in the Wait call. While
this is not necessarily harmful since the program is exiting, I would
feel better if I fixed it.