Ensure interval batcher always flushes output
A bug was encountered where the interval batcher would not always flush
its output. It was determined that this was due to the upstream batcher
emitting a nil value on its error channel before closing its output
channel, and the interval batcher not handling this case correctly.

The fix is to ensure that the interval batcher's goroutine defers
flusing of the output, such that all return paths are handled. In
addition, the interval batcher takes care to close both of its channels
(to ensure correct behavior if a client reads from the error channel but
the upstream batcher did not emit an error).

Where necessary, documentation is updated to reflect this change. Also,
additional color was added to the LineBatcher documentation regarding
behavior on EOF.