Add unit tests for Writer
Add the last set of basic unit tests for slackio. This set of tests, in
particular, deals with the Writer and covers most aspects of it. The
only parts missing coverage are the default batcher setup, and a panic
that I would like to think is relatively rare but that realistically I
should probably get rid of at some point.