Create GitLab CI configuration for tests
Set up a CI configuration that runs tests on the slackio package. This
is based on the slack-weather configuration, and includes the logic to
create a temporary $GOPATH in which the tests will be run. (This is kind
of a hack, ideally I'd run in Docker if I weren't so cheap with my VPS
hosting.)