Pass Message values, not pointers
While writing tests for slackio, I found that passing around references
to shared memory was surprisingly hard. Or, in other words, I made a lot
of stupid mistakes that I apparently didn't think through hard enough
and made strange behavior easy to create. For example, while ranging
over some stuff in a for loop I'd pass a pointer through a channel, only
to have the receiver on that channel read an unexpected value because
the loop variable had already changed after the receive operation.

The point is, I'm getting rid of pointers in the slackio API and am
passing values instead. Any efficiency hit is likely minimal,
realistically speaking, and the amount of mistakes this can potentially
reduce makes it well worth it in any case.