Implement async non-blocking subscriptions
This commit introduces a major rewrite of the slackio subscription
logic, which makes consumers of the message stream more independent
(such that a single consumer no longer blocks others) and removes all
potential deadlocks within the subscription management operations.

The downside of this change is that it's now possible for a consumer to
have dropped messages. Also, the subscription logic is way more complex.
But overall I'm still convinced this is beneficial.

It took a significant amount of careful thought to get to this
implementation, and I'm pleased to say that the stream and exec modes
appeared to work correctly on the first try once I finished it.
Nonetheless, more extensive testing will be needed before I'm fully
comfortable implementing the mux mode. At the moment the statement
coverage level for slackio has dropped significantly.