Snapshot first draft of childproc logic
This is a draft version of the childproc code that is really not that
good. There's a lot of weird closing and error handling going on that I
think can be simplified a good deal. In particular, I want copy
goroutines themselves to start owning more of the pipe closure logic
themselves. Then I can just wrap them all up in a WaitGroup and
termination becomes a lot easier to think about. I hope.