Snapshot second draft of childproc logic
This is the version that I wrote while waiting for some flights, with a
little bit done after nearly 24 hours of being awake and taking a short
break from catching up with famliy and some more done the next morning.
So... it's probably not 100% correct but there are aspects I like better
than before nonetheless.