Update and clean up childproc documentation
This includes various changes to clean things up a bit and hopefully
clarify a few smaller intricacies. Most notably, the Process.Wait method
is now documented, along with expectations for the io.ReadCloser given
to Spawn. The EPIPE logic was added as a TODO comment, since the whole
output draining thing feels silly.