Correctly set argv in childproc.Spawn
This fixes an EXTREMELY CRITICAL FLAW in childproc.Spawn that occurred
due to my misunderstanding of the os.StartProcess method.

StartProcess takes three arguments: the path of the executable to start,
the command line arguments as a []string, and a *ProcAttr for various
other attributes (e.g. open files that should be passed to the spawned
process).

When implementing the Spawn logic, I stupidly and incorrectly assumed
that the second parameter - the argv array - only needed to include
elements beyond argv[0] (the program name). This is roughly the
semantics of exec.Command from the os/exec package. However, the argv
array given to StartProcess is actually the EXACT argv array that the
child will receive. (Note that exec.Command handles this transparently
by appending the "arguments" portion to the passed-in name.)

The fix for this is to pass the entire "cmdline" given to Spawn as the
child's argv. This means that argv[0] for the child will be the string
that was passed to the LookPath call, i.e. this matches the semantics of
running a program from a shell.

The root cause of this critical flaw being introduced into slackbridge
is that I'm an idiot who fails to read documentation, fails to respect
the power of low-level interfaces, fails to THINK through the full
consequences of my code, fails to test my assertions properly before
trusting them (in this case by leveraging the broken childproc for the
exec subcommand), and fails to apply common sense (the fact that the
StartProcess parameter is named "argv" would have made this patently
obvious if I didn't suck).