Remove vendored dependencies
By transitioning to GitHub as the official home of slackbridge, I can
use Travis CI and get a real $GOPATH inside of a container, rather than
the hacky symlink-based setup I use on my GitLab server (that doesn't
work with `dep`).

The point of this is to let slackio be pulled in as an external package
if desired.