Disable cgo on Travis CI builds
Since Travis containers are linux/amd64 already (it seems), Go enables
cgo by default and links the resulting binary to libc. This results in
the binaries failing to start within Alpine Linux containers.

Upon testing with a temporary Docker build as well as analysis with
`ldd`, I can confirm that this should fix the issue.