Link MPL-licensed HashiCorp libraries in README
This README change is intended to help satisfy the terms of the Mozilla
Public License, version 2.0, under which the go-multierror and errwrap
packages are distributed.

In addition to this README change, slackbridge itself will be updated to
include information about the source code availability of these
libraries in its compiled releases.