Upgrade to Go 1.10
Of course, this brings in whatever performance and build improvements
were released in the new version.

(Fun fact: I was stupid the first time I did this, and wrote 1.10
without quotes. Of course, YAML parsed that as a float and I ended up
with Go 1.1 instead. On one hand, I should have recognized this, but on
the other hand... I'm surprised Travis doesn't enforce the data type
harder than this.)